// client/src/app/models/product.ts
export interface Product {
    
    
    id: number;
    name: string;
    description: string;
    price: number;
    pictureUrl: string;
    productBrand  :string ; // Assurez-vous que Brand est bien importé et défini correctement
    productType: string;   // Assurez-vous que Type est bien importé et défini correctement
}